rofl-oracle/requirements.txt
web3>=6.0.0
oasis-sapphire-py>=0.4.1
requests>=2.31.0
python-dotenv>=1.0.0
